reorganizing structure to accommodate Lambda layers
